+++ packages/pages/src/common/src/feature/stream.ts
@@ -22,7 +22,7 @@ export interface StreamConfig {
   /** The fields to apply to the stream */
   fields: string[];
   /** The filter to apply to the stream */
-  filter: {
+  filter?: {
     /** The entity IDs to apply to the stream */
     entityIds?: string[];
     /** The entity types to apply to the stream */
@@ -88,6 +88,15 @@ export const convertTemplateConfigToStreamConfig = (
   if (!config) {
     return;
   }
+
+  if (
+    config.stream &&
+    !config.stream.filter &&
+    !config.additionalProperties?.isVETemplate
+  ) {
+    logErrorAndExit("Filter is required in StreamConfig for templates.");
+  }
+
   if (config.stream) {
     return {
       ...config.stream,