+++ packages/pages/src/util/viteConfig.ts
@@ -16,10 +16,3 @@ export const scopedViteConfigPath = (scope?: string) => {
     return viteConfigPath;
   }
 };
-
-export const removePluginFromViteConfig = (config: any) => {
-  if (config?.plugins) {
-    config.plugins = "";
-  }
-  return config;
-};