+++ packages/pages/src/vite-plugin/modules/plugin.ts
@@ -1,4 +1,4 @@
-import { build, mergeConfig, Plugin } from "vite";
+import { build, mergeConfig, Plugin, UserConfig } from "vite";
 import { ProjectStructure } from "../../common/src/project/structure.js";
 import { glob } from "glob";
 import path from "node:path";
@@ -15,10 +15,7 @@ import postcss from "postcss";
 import nested from "postcss-nested";
 import { createModuleLogger } from "../../common/src/module/internal/logger.js";
 import { getModuleName } from "../../common/src/module/internal/getModuleConfig.js";
-import {
-  removePluginFromViteConfig,
-  scopedViteConfigPath,
-} from "../../util/viteConfig.js";
+import { scopedViteConfigPath } from "../../util/viteConfig.js";
 
 type FileInfo = {
   path: string;
@@ -66,7 +63,10 @@ export const buildModules = async (
   }
 
   const viteConfigPath = scopedViteConfigPath(projectStructure.config.scope);
-  const viteConfig = viteConfigPath ? await import(viteConfigPath) : "";
+  const viteConfigModule = viteConfigPath ? await import(viteConfigPath) : "";
+  const viteConfig = viteConfigModule
+    ? (viteConfigModule.default as UserConfig)
+    : undefined;
 
   for (const [moduleName, fileInfo] of Object.entries(filepaths)) {
     logger.info = (msg, options) => {
@@ -146,11 +146,9 @@ export const buildModules = async (
         }),
       ],
     };
+
     await build(
-      mergeConfig(
-        removePluginFromViteConfig(viteConfig.default),
-        moduleBuildConfig
-      )
+      mergeConfig(cleanseUserViteConfig(viteConfig), moduleBuildConfig)
     );
   }
 };
@@ -287,4 +285,20 @@ const tailwindBaseExists = (): boolean => {
     }
   }
   return false;
+};
+
+/**
+ * Removes items from the user-defined vite config that clash with the module build process.
+ */
+const cleanseUserViteConfig = (config?: UserConfig) => {
+  if (!config) {
+    return {};
+  }
+  config.plugins = [];
+
+  if (config.build?.rollupOptions?.output) {
+    config.build.rollupOptions.output = undefined;
+  }
+
+  return config;
 };