+++ packages/pages/src/vite-plugin/serverless-functions/plugin.ts
@@ -1,4 +1,10 @@
-import { build, createLogger, mergeConfig } from "vite";
+import {
+  build,
+  createLogger,
+  InlineConfig,
+  mergeConfig,
+  UserConfig,
+} from "vite";
 import { ProjectStructure } from "../../common/src/project/structure.js";
 import { glob } from "glob";
 import path from "node:path";
@@ -8,10 +14,7 @@ import { processEnvVariables } from "../../util/processEnvVariables.js";
 import { FunctionMetadataParser } from "../../common/src/function/internal/functionMetadataParser.js";
 import { nodePolyfills } from "vite-plugin-node-polyfills";
 import pc from "picocolors";
-import {
-  removePluginFromViteConfig,
-  scopedViteConfigPath,
-} from "../../util/viteConfig.js";
+import { scopedViteConfigPath } from "../../util/viteConfig.js";
 
 export const buildServerlessFunctions = async (
   projectStructure: ProjectStructure
@@ -46,7 +49,10 @@ export const buildServerlessFunctions = async (
   const loggerInfo = logger.info;
 
   const viteConfigPath = scopedViteConfigPath(projectStructure.config.scope);
-  const viteConfig = viteConfigPath ? await import(viteConfigPath) : "";
+  const viteConfigModule = viteConfigPath ? await import(viteConfigPath) : "";
+  const viteConfig = viteConfigModule
+    ? (viteConfigModule.default as UserConfig)
+    : undefined;
 
   for (const [name, filepath] of Object.entries(filepaths)) {
     logger.info = (msg, options) => {
@@ -60,7 +66,7 @@ export const buildServerlessFunctions = async (
       loggerInfo(msg, options);
     };
 
-    const serverlessFunctionBuildConfig = {
+    const serverlessFunctionBuildConfig: InlineConfig = {
       customLogger: logger,
       configFile: false,
       envDir: envVarConfig.envVarDir,
@@ -96,12 +102,21 @@ export const buildServerlessFunctions = async (
         }),
       ],
     };
-    await build(
-      mergeConfig(
-        removePluginFromViteConfig(viteConfig.default),
-        serverlessFunctionBuildConfig
-      )
-    );
+
+    const mergedConfig = viteConfig?.build?.rollupOptions?.external
+      ? mergeConfig(
+          {
+            build: {
+              rollupOptions: {
+                external: viteConfig.build.rollupOptions.external,
+              },
+            },
+          },
+          serverlessFunctionBuildConfig
+        )
+      : serverlessFunctionBuildConfig;
+
+    await build(mergedConfig);
   }
 };